Move exported enums
