+++ dashboard/src/components/AppView/AppView.test.tsx
@@ -28,8 +28,7 @@ import { IConfigState } from "reducers/config";
 import { InstalledPackage } from "shared/InstalledPackage";
 import PackagesService from "shared/PackagesService";
 import { getStore, mountWrapper } from "shared/specs/mountWrapper";
-import { DeleteError, FetchError, IInstalledPackageState, IStoreState } from "shared/types";
-import { PluginNames } from "shared/utils";
+import { DeleteError, FetchError, IInstalledPackageState, IStoreState, PluginNames } from "shared/types";
 import { getType } from "typesafe-actions";
 import AccessURLTable from "./AccessURLTable/AccessURLTable";
 import DeleteButton from "./AppControls/DeleteButton/DeleteButton";
