+++ dashboard/src/components/Catalog/CatalogItems.test.tsx
@@ -5,8 +5,7 @@ import InfoCard from "components/InfoCard";
 import { AvailablePackageSummary, Context } from "gen/kubeappsapis/core/packages/v1alpha1/packages";
 import { Plugin } from "gen/kubeappsapis/core/plugins/v1alpha1/plugins";
 import { defaultStore, mountWrapper } from "shared/specs/mountWrapper";
-import { IClusterServiceVersion } from "shared/types";
-import { PluginNames } from "shared/utils";
+import { IClusterServiceVersion, PluginNames } from "shared/types";
 import CatalogItem from "./CatalogItem";
 import CatalogItems, { ICatalogItemsProps } from "./CatalogItems";
